Feature/test navigation transition (#4)
* [#3] MainPageScreen -> MainScreenCore 폴더이름 변경

* [#3] MainCoordinator Core, View 파일 분리
